+++ GetThePicture/Picture/Clause/Base/Computational/PackedDecimal.cs
@@ -1,5 +1,6 @@
 using GetThePicture.Picture.Clause.Base.Options;
-using GetThePicture.Picture.Clause.Encoder.Category;
+using GetThePicture.Picture.Clause.Codec.Category.Numeric;
+using GetThePicture.Picture.Clause.Codec.Category.Numeric.Mapper;
 
 namespace GetThePicture.Picture.Clause.Base.Computational;
 
@@ -50,22 +51,25 @@ internal static class COMP3
     // - Total bytes = (number_of_digits + 1) / 2
     //
 
+    private static readonly SIntMapper _SIntMapper = new();
+    private static readonly UIntMapper _UIntMapper = new();
+
     public static object Decode(ReadOnlySpan<byte> buffer, PicMeta pic, DataStorageOptions ds = DataStorageOptions.CI)
     {
+        // Decode BCD
         byte[] chars = DecodePacked(buffer, pic.DigitCount, out bool isNegative); // 根據 PIC 長度解碼 BCD
 
-        if (pic.DecimalDigits > 0)
-            return CbDecimal.Decode(chars, pic.DecimalDigits, isNegative);
-        
-        if (!pic.Signed)
-        {
-            if (isNegative)
-                throw new OverflowException("Unsigned field contains negative number");
-                
-            return DecodeUInt64(chars);
-        }
+        if (!pic.Signed && isNegative)
+            throw new OverflowException("Unsigned field contains negative number");
 
-        return DecodeInt64(chars, isNegative);
+        decimal value = CbDecimal.Decode(chars, pic.DecimalDigits, isNegative);
+
+        if (pic.DecimalDigits > 0)
+            return value;
+
+        IMapper mapper = pic.Signed ? _SIntMapper : _UIntMapper;
+
+        return mapper.Map(value, pic);
     }
 
     public static byte[] Encode(NumericMeta nMeta, PicMeta pic, DataStorageOptions ds = DataStorageOptions.CI)
@@ -99,6 +103,9 @@ internal static class COMP3
 
     private static byte[] DecodePacked(ReadOnlySpan<byte> buffer, int digits, out bool negative)
     {
+        if (digits < 1)
+            throw new ArgumentOutOfRangeException(nameof(digits), "Digits must be greater than 0.");
+        
         byte[] bytes = new byte[digits];
 
         int idx = digits - 1;
@@ -135,35 +142,5 @@ internal static class COMP3
         }
 
         return bytes;
-    }
-
-    private static long DecodeInt64(ReadOnlySpan<byte> chars, bool isNegative)
-    {
-        if (chars.Length > 18)
-            throw new OverflowException("Packed number too large for Int64");
-
-        long value = 0;
-        foreach (byte c in chars)
-            value = value * 10 + (c - '0');
-
-        return isNegative ? -value : value;
-    }
-
-    private static ulong DecodeUInt64(ReadOnlySpan<byte> chars)
-    {
-        ulong value = 0;
-
-        foreach (byte c in chars)
-        {
-            ulong digit = (ulong)(c - '0');
-
-            // overflow check
-            if (value > (ulong.MaxValue - digit) / 10)
-                throw new OverflowException("Packed number too large for UInt64");
-
-            value = value * 10 + digit;
-        }
-
-        return value;
     }
 }